add: task削除エンドポイントを作成

+++ ApolloServer.js
@@ -22,13 +22,30 @@ const resolvers = {
             return newTask;
         },
         deleteTask: async (parent, args, context) => {
-            console.log(args.id);
+            const task = await prisma.task.findUnique({
+                where: {
+                    id: args.id
+                },
+            });
+            if (!task) {
+                return {
+                    errors: [
+                        {
+                            message: "削除対象のタスクがありません",
+                        }
+                    ],
+                    task: null,
+                };
+            }
             await context.prisma.task.delete({
                 where: {
                     id: args.id
                 }
             });
-            return true;
+            return {
+                errors: [],
+                task,
+            };
         }
     }
 };
